requirements.txt
fastapi==0.79.0
numpy==1.23.1
opencv_python==4.6.0.66
Pillow==9.2.0
streamlit==1.11.1
